Remove explicitly set line-height values
We were explicitly setting line-heights on the button and the toggle - normally the
govuk-font mixin will do this for us and will choose a line-height value appropriate to
the font size.

We were seeing an issue with the focus state on smaller screens where there was
a thin black line between lines. See https://github.com/alphagov/govuk-frontend/issues/2322
 for more detail.

This commit removes the explicitly set line-height so that the focus state displays as
expected.