Remove accordion <sup> examples
These examples were added to test a different combination of HTML elements in the
accordion component. However, the <sup> element specifically pushes text in the
accordion out of line and therefore requires additional styling if it's going to be
used (vertical-align: top). We also don't think it's likely that a user would want
to use <sup> inside the summary line.

This commit removes the <sup> from the accordion examples. The existing examples still
test the accordion when passing in HTML content, so that is still covered.