Fix focus state gap on desktop
Fixes https://github.com/alphagov/govuk-frontend/issues/2371

This commit removes the font and line-height being set on the accordion button, which we were relying on to style the accordion heading text. It seemed like this was affecting the other text elements inside the accordion sections and impacting the focus state.

Instead, we are directly setting the font and line height on the accordion heading text. Although browsers will set their own
font styles for buttons, this shouldn't affect the component because we're setting the text styles on the individual span elements within the button.
These should override any browser styles being set.

We also need to adjust the spacing between text elements a bit as a result, otherwise the height of the accordion (and therefore the touch target
is reduced). Read more about that here https://github.com/alphagov/govuk-frontend/issues/2371#issuecomment-932126039